Struct owapi::v3::achievements::OffenseAchievements

pub struct OffenseAchievements {
    pub hack_the_planet: bool,
    pub power_outage: bool,
    pub whoa_there: bool,
    pub die_die_die_die: bool,
    pub its_high_noon: bool,
    pub their_own_worst_enemy: bool,
    pub clearing_the_area: bool,
    pub target_rich_environment: bool,
    pub death_from_above: bool,
    pub total_recall: bool,
    pub rocket_man: bool,
    pub slice_and_dice: bool,
    pub special_delivery: bool,
    pub waste_not_want_not: bool,
}

Achievements for heroes classified by Blizzard as "offense".

Fields
hack_the_planet: bool

Hack The Planet achievement.

Hack 15 enemies without dying as Sombra in quick or competitive play.

power_outage: bool

Power Outage achievement.

Hack 6 enemies at once as Sombra in quick or competitive play.

whoa_there: bool

Whoa There! achievement.

Interrupt an enemy ultimate ability with McCree's Flashbang.

die_die_die_die: bool

Die Die Die... Die achievement.

Kill 4 enemies with a single use of Reaper's Death Blossom.

its_high_noon: bool

It's High Noon achievement.

Get 4 killing blows with a single use of McCree's Deadeye.

their_own_worst_enemy: bool

Their Own Worst Enemy achievement.

Kill 2 enemies with a single use of Genji's Deflect.

clearing_the_area: bool

Clearing the Area achievement.

Knock an enemy to their death using Pharah's Concussive Blast.

target_rich_environment: bool

Target Rich Environment achievement.

Kill 4 enemies with a single use of Soldier: 76's Tactical Visor.

death_from_above: bool

Death From Above achievement.

Kill 4 enemies in a row without touching the ground as Pharah.

total_recall: bool

Total Recall achievement.

Recover 400 health using Tracer's Recall without dying.

rocket_man: bool

Rocket Man achievement.

Get 2 killing blows with a single use of Soldier: 76's Helix Rockets.

slice_and_dice: bool

Slice and Dice achievement.

Kill 4 enemies with a single use of Genji's Dragonblade.

special_delivery: bool

Special Delivery achievement.

Stick 4 of Tracer's Pulse Bombs onto enemies in a single game.

waste_not_want_not: bool

Waste Not, Want Not achievement.

Get 3 solo kills with a single clip of Reaper's shotguns.