Struct owapi::v3::achievements::SpecialAchievements

pub struct SpecialAchievements {
    pub survived_the_night: bool,
    pub flagbearer: bool,
    pub cleanup_duty: bool,
    pub not_a_scratch: bool,
    pub four_they_were: bool,
    pub whap: bool,
    pub cool_as_ice: bool,
    pub ambush: bool,
    pub snowed_in: bool,
    pub held_the_door: bool,
}

Achievements for various events.

Fields
survived_the_night: bool

Survived The Night achievement.

Win Junkenstein's Revenge on Medium difficulty.

flagbearer: bool

Flagbearer achievement.

Win Capture the Rooster with as score of 3 to 0.

cleanup_duty: bool

Cleanup Duty achievement.

Recover a flag in a game of Capture the Rooster.

not_a_scratch: bool

Not A Scratch achievement.

Win Junkenstein's Revenge on Medium difficulty with no damage to the door.

four_they_were: bool

Four They Were achievement.

Win Junkenstein's Revenge using each of the 4 heroes.

whap: bool

Whap! achievement.

Kill an enemy with a snowball from 25 meters away in Mei's Snowball Offensive.

cool_as_ice: bool

Cool As Ice achievement.

Kill 4 enemies without missing in Mei's Snowball Offensive.

ambush: bool

Ambush! achievement.

Kill 3 enemies while they are picking up snow in a game of Mei's Snowball Offensive.

snowed_in: bool

Snowed In achievement.

Win Mei's Snowball Offensive without losing a round.

held_the_door: bool

Held The Door achievement.

Win Junkenstein's Revenge on Hard difficulty.