Struct owapi::v3::stats::OverallStats

pub struct OverallStats {
    pub win_rate: f64,
    pub level: u64,
    pub prestige: u64,
    pub avatar: String,
    pub rank_image: Option<String>,
    pub tier: Option<String>,
    pub wins: u64,
    pub games: u64,
    pub comprank: u64,
    pub losses: u64,
    pub ties: Option<u64>,
}

Overall statistics.

Fields
win_rate: f64

% of games won.

level: u64

User's level, excl. prestige ("stars")

The user's numerical level can be calculated using the (prestige * 100) + level formula.

prestige: u64

User's prestige ("stars")

The user's numerical level can be calculated using the (prestige * 100) + level formula.

avatar: String

URL to user's avatar (little chooseable image, e.g. a pachimari).

rank_image: Option<String>

URL to user's portrait.

tier: Option<String>

User's competitive "tier", like "gold" or "bronze".

wins: u64

Amount of games won overall.

games: u64

Amount of games played overall.

comprank: u64

User's SR.

losses: u64

Amount of games lost overall.

ties: Option<u64>

Amount of games tied overall.